requirements.txt
django
Pillow
django-filter
rest_framework
flake8